refactor: reorganize compiler integration test modules
