+++ tests/integration/src/codegen/control_flow/mod.rs
@@ -1,0 +1,2 @@
+mod select;
+mod switch;
