+++ tests/integration/src/codegen/control_flow/select.rs
@@ -176,5 +176,3 @@ fn select_i128() {
         ],
     );
 }
-
-mod switch;
