--- tests/integration/src/codegen/intrinsics/mod.rs
@@ -1,1 +1,0 @@
-mod mem;
