+++ tests/integration/src/codegen/memory/mod.rs
@@ -1,0 +1,1 @@
+mod property;
