+++ tests/integration/src/codegen/memory/property.rs
