+++ tests/integration/src/codegen/wasm/functional.rs
