+++ tests/integration/src/codegen/wasm/mod.rs
@@ -1,0 +1,1 @@
+mod functional;
