+++ tests/integration/src/rust_pipeline/abi_transform/stdlib.rs
