+++ tests/integration/src/rust_pipeline/apps.rs
