+++ tests/integration/src/rust_pipeline/debug_source_locations.rs
