+++ tests/integration/src/rust_pipeline/examples.rs
