+++ tests/integration/src/rust_pipeline/instructions.rs
