+++ tests/integration/src/rust_pipeline/misc.rs
