+++ tests/integration/src/rust_pipeline/mod.rs
