+++ tests/integration/src/rust_pipeline/rust_sdk/base/faucet.rs
