+++ tests/integration/src/rust_pipeline/rust_sdk/base/input_note.rs
