+++ tests/integration/src/rust_pipeline/rust_sdk/base/mod.rs
