+++ tests/integration/src/rust_pipeline/rust_sdk/base/note.rs
