+++ tests/integration/src/rust_pipeline/rust_sdk/base/output_note.rs
