+++ tests/integration/src/rust_pipeline/rust_sdk/stdlib/collections.rs
