+++ tests/integration/src/rust_pipeline/rust_sdk/stdlib/hashes.rs
