+++ tests/integration/src/rust_pipeline/types.rs
