+++ tests/integration/src/rust_pipeline/types_src/enum.rs
