+++ tests/integration/src/rust_pipeline/types_src/static_mut.rs
